pyproject.toml
[project]
name = "indigo-assignment"
version = "0.1.0"
description = "Knowledge base backend with FastAPI, Postgres/pgvector, and an MCP server over Streamable HTTP."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.3",
  "pgvector>=0.3.6",
  "pydantic-settings>=2.6.1",
  "python-multipart>=0.0.12",
  "openai>=1.54.0",
  "pypdf>=5.1.0",
  "mcp[cli]>=1.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "httpx>=0.27.2",
  "ruff>=0.7.2",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
